Separate all steps to files
