+++ query/linkedql/iter_tags.go
@@ -13,27 +13,27 @@ var _ query.Iterator = (*TagsIterator)(nil)
 // TagsIterator is a result iterator for records consisting of selected tags
 // or all the tags in the query.
 type TagsIterator struct {
-	valueIt  *ValueIterator
-	selected []string
+	ValueIt  *ValueIterator
+	Selected []string
 }
 
 // Next implements query.Iterator.
 func (it *TagsIterator) Next(ctx context.Context) bool {
-	return it.valueIt.Next(ctx)
+	return it.ValueIt.Next(ctx)
 }
 
 func (it *TagsIterator) getTags() map[string]interface{} {
 	refTags := make(map[string]refs.Ref)
-	it.valueIt.scanner.TagResults(refTags)
+	it.ValueIt.scanner.TagResults(refTags)
 
 	tags := make(map[string]interface{})
-	if it.selected != nil {
-		for _, tag := range it.selected {
-			tags[tag] = jsonld.FromValue(it.valueIt.getName(refTags[tag]))
+	if it.Selected != nil {
+		for _, tag := range it.Selected {
+			tags[tag] = jsonld.FromValue(it.ValueIt.getName(refTags[tag]))
 		}
 	} else {
 		for tag, ref := range refTags {
-			tags[tag] = jsonld.FromValue(it.valueIt.getName(ref))
+			tags[tag] = jsonld.FromValue(it.ValueIt.getName(ref))
 		}
 	}
 
@@ -47,10 +47,10 @@ func (it *TagsIterator) Result() interface{} {
 
 // Err implements query.Iterator.
 func (it *TagsIterator) Err() error {
-	return it.valueIt.Err()
+	return it.ValueIt.Err()
 }
 
 // Close implements query.Iterator.
 func (it *TagsIterator) Close() error {
-	return it.valueIt.Close()
+	return it.ValueIt.Close()
 }
