+++ query/linkedql/steps/back.go
@@ -1,0 +1,47 @@
+package steps
+
+import (
+	"github.com/cayleygraph/cayley/graph"
+	"github.com/cayleygraph/cayley/query"
+	"github.com/cayleygraph/cayley/query/linkedql"
+	"github.com/cayleygraph/cayley/query/path"
+	"github.com/cayleygraph/quad"
+	"github.com/cayleygraph/quad/voc"
+)
+
+func init() {
+	linkedql.Register(&Back{})
+}
+
+var _ linkedql.IteratorStep = (*Back)(nil)
+var _ linkedql.PathStep = (*Back)(nil)
+
+// Back corresponds to .back().
+type Back struct {
+	From linkedql.PathStep `json:"from"`
+	Name string            `json:"name"`
+}
+
+// Type implements Step.
+func (s *Back) Type() quad.IRI {
+	return linkedql.Prefix + "Back"
+}
+
+// Description implements Step.
+func (s *Back) Description() string {
+	return "resolves to the values of the previous the step or the values assigned to name in a former step."
+}
+
+// BuildIterator implements linkedql.IteratorStep.
+func (s *Back) BuildIterator(qs graph.QuadStore, ns *voc.Namespaces) (query.Iterator, error) {
+	return linkedql.NewValueIteratorFromPathStep(s, qs, ns)
+}
+
+// BuildPath implements linkedql.PathStep.
+func (s *Back) BuildPath(qs graph.QuadStore, ns *voc.Namespaces) (*path.Path, error) {
+	fromPath, err := s.From.BuildPath(qs, ns)
+	if err != nil {
+		return nil, err
+	}
+	return fromPath.Back(s.Name), nil
+}
