+++ query/linkedql/steps/greater_than.go
@@ -1,0 +1,50 @@
+
+package steps
+
+import (
+	"github.com/cayleygraph/cayley/graph"
+	"github.com/cayleygraph/cayley/graph/iterator"
+	"github.com/cayleygraph/cayley/query"
+	"github.com/cayleygraph/cayley/query/path"
+	"github.com/cayleygraph/quad"
+	"github.com/cayleygraph/quad/voc"
+	"github.com/cayleygraph/cayley/query/linkedql"
+)
+
+func init() {
+	linkedql.Register(&GreaterThan{})
+}
+
+var _ linkedql.IteratorStep = (*GreaterThan)(nil)
+var _ linkedql.PathStep = (*GreaterThan)(nil)
+
+// GreaterThan corresponds to gt().
+type GreaterThan struct {
+	From  linkedql.PathStep   `json:"from"`
+	Value quad.Value `json:"value"`
+}
+
+// Type implements Step.
+func (s *GreaterThan) Type() quad.IRI {
+	return linkedql.Prefix + "GreaterThan"
+}
+
+// Description implements Step.
+func (s *GreaterThan) Description() string {
+	return "Greater than equals filters out values that are not greater than given value"
+}
+
+// BuildIterator implements linkedql.IteratorStep.
+func (s *GreaterThan) BuildIterator(qs graph.QuadStore, ns *voc.Namespaces) (query.Iterator, error) {
+	return linkedql.NewValueIteratorFromPathStep(s, qs, ns)
+}
+
+// BuildPath implements linkedql.PathStep.
+func (s *GreaterThan) BuildPath(qs graph.QuadStore, ns *voc.Namespaces) (*path.Path, error) {
+	fromPath, err := s.From.BuildPath(qs, ns)
+	if err != nil {
+		return nil, err
+	}
+	return fromPath.Filter(iterator.CompareGT, s.Value), nil
+}
+
