+++ query/linkedql/steps/property_names.go
@@ -1,0 +1,47 @@
+package steps
+
+import (
+	"github.com/cayleygraph/cayley/graph"
+	"github.com/cayleygraph/cayley/query"
+	"github.com/cayleygraph/cayley/query/path"
+	"github.com/cayleygraph/quad"
+	"github.com/cayleygraph/quad/voc"
+	"github.com/cayleygraph/cayley/query/linkedql"
+)
+
+func init() {
+	linkedql.Register(&PropertyNames{})
+}
+
+var _ linkedql.IteratorStep = (*PropertyNames)(nil)
+var _ linkedql.PathStep = (*PropertyNames)(nil)
+
+// PropertyNames corresponds to .propertyNames().
+type PropertyNames struct {
+	From linkedql.PathStep `json:"from"`
+}
+
+// Type implements Step.
+func (s *PropertyNames) Type() quad.IRI {
+	return linkedql.Prefix + "PropertyNames"
+}
+
+// Description implements Step.
+func (s *PropertyNames) Description() string {
+	return "gets the list of predicates that are pointing out from a node."
+}
+
+// BuildIterator implements linkedql.IteratorStep.
+func (s *PropertyNames) BuildIterator(qs graph.QuadStore, ns *voc.Namespaces) (query.Iterator, error) {
+	return linkedql.NewValueIteratorFromPathStep(s, qs, ns)
+}
+
+// BuildPath implements linkedql.PathStep.
+func (s *PropertyNames) BuildPath(qs graph.QuadStore, ns *voc.Namespaces) (*path.Path, error) {
+	fromPath, err := s.From.BuildPath(qs, ns)
+	if err != nil {
+		return nil, err
+	}
+	return fromPath.OutPredicates(), nil
+}
+
