+++ query/linkedql/steps/skip.go
@@ -1,0 +1,48 @@
+package steps
+
+import (
+	"github.com/cayleygraph/cayley/graph"
+	"github.com/cayleygraph/cayley/query"
+	"github.com/cayleygraph/cayley/query/path"
+	"github.com/cayleygraph/quad"
+	"github.com/cayleygraph/quad/voc"
+	"github.com/cayleygraph/cayley/query/linkedql"
+)
+
+func init() {
+	linkedql.Register(&Skip{})
+}
+
+var _ linkedql.IteratorStep = (*Skip)(nil)
+var _ linkedql.PathStep = (*Skip)(nil)
+
+// Skip corresponds to .skip().
+type Skip struct {
+	From   linkedql.PathStep `json:"from"`
+	Offset int64    `json:"offset"`
+}
+
+// Type implements Step.
+func (s *Skip) Type() quad.IRI {
+	return linkedql.Prefix + "Skip"
+}
+
+// Description implements Step.
+func (s *Skip) Description() string {
+	return "skips a number of nodes for current path."
+}
+
+// BuildIterator implements linkedql.IteratorStep.
+func (s *Skip) BuildIterator(qs graph.QuadStore, ns *voc.Namespaces) (query.Iterator, error) {
+	return linkedql.NewValueIteratorFromPathStep(s, qs, ns)
+}
+
+// BuildPath implements linkedql.PathStep.
+func (s *Skip) BuildPath(qs graph.QuadStore, ns *voc.Namespaces) (*path.Path, error) {
+	fromPath, err := s.From.BuildPath(qs, ns)
+	if err != nil {
+		return nil, err
+	}
+	return fromPath.Skip(s.Offset), nil
+}
+
